refactor storage.h
